requirements.txt
gradio>=3.27.0
langchain>=0.0.151
marvin>=0.8.0
python-dotenv>=1.0.0